4v

+++ C++ projects/first_lab_version2/first_lab_version2/astar.h
@@ -19,7 +19,7 @@ public:
     float h_value(Node current, Node goal, Graph graph);
     struct CustomCompare
     {
-        bool operator()(const std::pair<int, Node*>& lhs, const std::pair<int, Node*>& rhs)
+        bool operator()(const std::pair<float, Node*>& lhs, const std::pair<float, Node*>& rhs)
         {
             return lhs.first > rhs.first;
         }
